+++ account_invoice_validate_send_email/i18n/ja.po
@@ -127,3 +127,7 @@ msgstr "ワークフロープロセス設定にかかわらず、この項目が
 msgid "【請求書】送付のご案内　${object.company_id.name}　${object.number or \"N/A\"}"
 msgstr "【請求書】送付のご案内　${object.company_id.name}　${object.number or \"N/A\"}"
 
+#. module: stock_delivery_result_import
+#: model:ir.model.fields,field_description:stock_delivery_result_import.field_account_invoice_yamato_slip_number
+msgid "Yamato Slip Number"
+msgstr "配送伝票番号"
