+++ account_invoice_validate_send_email/models/account_invoice.py
@@ -18,6 +18,12 @@ class AccountInvoice(models.Model):
     )
     web_url = fields.Char()
 
+    yamato_slip_number = fields.Text(
+        "Yamato Slip Number",
+        help="Delivery Slip Number of Yamato Transport CO.,.",
+        compute="_compute_yamato_slip_number",
+    )
+
     def _get_mail_template(self):
         self.ensure_one()
         return self.company_id.invoice_mail_template_id
@@ -109,3 +115,12 @@ class AccountInvoice(models.Model):
                 )
                 self.invoice_sent = True
         return True
+
+    @api.model
+    def _compute_yamato_slip_number(self):
+        for invoice in self:
+            tracking_numbers = ""
+            for picking in invoice.picking_ids:
+                if picking.carrier_tracking_ref:
+                    tracking_numbers += picking.carrier_tracking_ref + "\n"
+            invoice.yamato_slip_number = tracking_numbers
