+++ account_invoice_validate_send_email/views/account_invoice_view.xml
@@ -8,6 +8,7 @@
             <xpath expr="//field[@name='move_id']" position="after">
                 <field name="invoice_sent" />
                 <field name="web_url" invisible="1" />
+                <field name="yamato_slip_number" invisible="1"/>
             </xpath>
         </field>
     </record>
