+++ stock_delivery_result_import/__manifest__.py
@@ -8,16 +8,8 @@
     "license": "LGPL-3",
     "author": "Quartile Limited",
     "website": "https://www.quartile.co/",
-    "depends": [
-        "stock",
-        "base_data_import",
-        "queue_job",
-        "stock_picking_yamato_csv",
-        "stock_picking_invoice_link",
-        "account",
-    ],
+    "depends": ["stock", "base_data_import", "queue_job"],
     "data": [
-        "views/account_invoice_views.xml",
         "views/data_import_log_views.xml",
         "wizards/stock_delivery_result_import_views.xml",
     ],
