+++ stock_delivery_result_import/models/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import account_invoice
 from . import data_import_log
 from . import stock_picking
