--- stock_delivery_result_import/views/account_invoice_views.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="pro_mi7_invoice_supplier_form_inherit" model="ir.ui.view">
-        <field name="name">account.invoice.form</field>
-        <field name="model">account.invoice</field>
-        <field name="inherit_id" ref="account.invoice_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='date_due']" position="after">
-                <field name="yamato_slip_number" invisible="1"/>
-            </xpath>
-        </field>
-    </record>
-</odoo>
